update current useScaffoldEventHistory
